requirements.txt
langchain
langchain-community
langchain-openai
googlesearch-python
openai
langchainhub
requests
fastapi
dotenv
google-api-python-client
google-generativeai